bla bla

+++ routes/web.php
@@ -21,17 +21,19 @@ Route::get('/', function () {
 
 Route::get("/products", [ProductController::class, "index"]);
 
-Route::get("/products/create", [ProductController::class, "create"]);
-Route::post("/products", [ProductController::class, "store"]);
+Route::get("/products/create", [ProductController::class, "create"])->middleware("auth");
+Route::post("/products", [ProductController::class, "store"])->middleware("auth");
 Route::get("/products/{id}", [ProductController::class, "show"]);
-Route::get("/products/{id}/edit", [ProductController::class, "edit"]);
-Route::put("/products/{id}", [ProductController::class, "update"]);
-Route::delete("products/{id}", [ProductController::class, "destroy"]);
+Route::get("/products/{id}/edit", [ProductController::class, "edit"])->middleware("auth");
+Route::put("/products/{id}", [ProductController::class, "update"])->middleware("auth");
+Route::delete("products/{id}", [ProductController::class, "destroy"])->middleware("auth");
 
-Route::get("/register", [UserController::class, "create"]);
-Route::post("/register", [UserController::class, "store"]);
+Route::get("/register", [UserController::class, "create"])->middleware("guest");
+Route::post("/register", [UserController::class, "store"])->middleware("guest");
 
-Route::get("/login", [UserController::class, "login"]);
-Route::post("/login", [UserController::class, "signin"]); // Ļoti nepareizi, labāk būtu login2, vēl labāk - uztaisīt atsevišķu Controller
+Route::get("/login", [UserController::class, "login"])->name("login")->middleware("guest");
+Route::post("/login", [UserController::class, "signin"])->middleware("guest"); // Ļoti nepareizi, labāk būtu login2, vēl labāk - uztaisīt atsevišķu Controller
 
-Route::post("/logout",[UserController::class, "logout"]);
+Route::post("/logout",[UserController::class, "logout"])->middleware("auth");
+
+
